refactor: rename routes to controller
